backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
cryptography
python-jose
python-multipart
ratelimit
pyjwt